backend/requirements.txt
fastapi
uvicorn
transformers
huggingface_hub
openai
torch
numpy
python-dotenv